+++ Src/AwesomeAssertions/Equivalency/SelfReferenceEquivalencyOptions.cs
@@ -389,11 +389,13 @@ public abstract class SelfReferenceEquivalencyOptions<TSelf> : IEquivalencyOptio
     /// <summary>
     /// Excludes the members with the specified names from the structural equality check.
     /// </summary>
+    /// <param name="memberNames">Names of members to exclude.</param>
+    /// <exception cref="ArgumentNullException"><paramref name="memberNames"/> is <see langword="null"/>.</exception>
+    /// <exception cref="ArgumentException"><paramref name="memberNames"/> is empty.</exception>
     public TSelf ExcludingMembersNamed(params string[] memberNames)
     {
         Guard.ThrowIfArgumentIsNull(memberNames);
         Guard.ThrowIfArgumentIsEmpty(memberNames, nameof(memberNames), "At least one member name must be specified.");
-        Guard.ThrowIfArgumentContainsNullOrEmpty(memberNames);
 
         AddSelectionRule(new ExcludeMemberByNameSelectionRule(memberNames));
         return (TSelf)this;