+++ Tests/AwesomeAssertions.Equivalency.Specs/SelectionRulesSpecs.Excluding.cs
@@ -70,14 +70,14 @@ public partial class SelectionRulesSpecs
             {
                 Id = 1,
                 Name = "John",
-                Age = 13
+                Age = 13,
             };
 
             var customer = new
             {
                 Id = 2,
                 Name = "Jack",
-                Age = 37
+                Age = 37,
             };
 
             // Act
@@ -90,33 +90,53 @@ public partial class SelectionRulesSpecs
         }
 
         [Fact]
-        public void A_member_excluded_by_name_should_not_fail()
+        public void Excluding_member_by_name_should_succeed()
         {
             // Arrange
             var subject = new
             {
                 Id = 1,
                 Name = "John",
-                Age = 13
+                Age = 13,
             };
 
             var customer = new
             {
                 Id = 1,
                 Name = "Jack",
-                Age = 13
+                Age = 13,
+            };
+
+            // Act / Assert
+            subject.Should().BeEquivalentTo(customer, options => options
+                .ExcludingMembersNamed("Name"));
+        }
+
+        [Fact]
+        public void Excluding_member_by_name_respects_casing()
+        {
+            // Arrange
+            var subject = new
+            {
+                Name = "John",
+            };
+
+            var customer = new
+            {
+                Name = "Jack",
             };
 
             // Act
             Action act = () => subject.Should().BeEquivalentTo(customer, options => options
-                .ExcludingMembersNamed("Name"));
+                .ExcludingMembersNamed("name"));
 
             // Assert
-            act.Should().NotThrow();
+            act.Should().Throw<XunitException>()
+                .Which.Message.Should().Match("Expected property subject.Name*- Exclude members named: name*");
         }
 
         [Fact]
-        public void A_member_excluded_by_name_should_pass_in_nested_types()
+        public void Excluding_member_by_name_should_pass_in_nested_types()
         {
             // Arrange
             var subject = new
@@ -126,7 +146,7 @@ public partial class SelectionRulesSpecs
                 {
                     Name = "John",
                     Age = 13,
-                }
+                },
             };
 
             var customer = new
@@ -135,24 +155,21 @@ public partial class SelectionRulesSpecs
                 Nested = new
                 {
                     Age = 13,
-                }
+                },
             };
 
-            // Act
-            Action act = () => subject.Should().BeEquivalentTo(customer, options => options
+            // Act / Assert
+            subject.Should().BeEquivalentTo(customer, options => options
                 .ExcludingMembersNamed("Name"));
-
-            // Assert
-            act.Should().NotThrow();
         }
 
         [Fact]
-        public void When_a_null_parameter_is_passed_instead_of_memeber_names_it_should_throw()
+        public void Passing_a_null_parameter_as_member_names_should_throw()
         {
             // Arrange
             var subject = new
             {
-                Age = 10
+                Age = 10,
             };
 
             // Act
@@ -165,12 +182,12 @@ public partial class SelectionRulesSpecs
         }
 
         [Fact]
-        public void When_an_empty_collection_parameter_is_passed_instead_of_memeber_names_it_should_throw()
+        public void Passing_an_empty_collection_parameter_as_member_names_should_throw()
         {
             // Arrange
             var subject = new
             {
-                Age = 10
+                Age = 10,
             };
 
             // Act
@@ -180,26 +197,7 @@ public partial class SelectionRulesSpecs
             // Assert
             act.Should().Throw<ArgumentException>()
                 .WithParameterName("memberNames")
-                .WithMessage("At least one member name must be specified.\nParameter name: memberNames");
-        }
-
-        [Fact]
-        public void When_an_empty_or_null_member_name_is_passed_in_memeber_names_parameter_it_should_throw()
-        {
-            // Arrange
-            var subject = new
-            {
-                Age = 10
-            };
-
-            // Act
-            Action act = () => subject.Should().BeEquivalentTo(subject, options => options
-                .ExcludingMembersNamed([null, "Age"]));
-
-            // Assert
-            act.Should().Throw<ArgumentException>()
-                .WithParameterName("memberNames")
-                .WithMessage("Collection contains a null or empty value\nParameter name: memberNames");
+                .WithMessage("At least one member name must be specified*memberNames*");
         }
 
         [Fact]
@@ -269,14 +267,9 @@ public partial class SelectionRulesSpecs
                 Field2 = "ipsum"
             };
 
-            // Act
-            Action act =
-                () =>
-                    class1.Should().BeEquivalentTo(class2,
-                        opts => opts.Excluding(o => o.Field3).Excluding(o => o.Property1));
-
-            // Assert
-            act.Should().NotThrow("the non-excluded fields have the same value");
+            // Act / Assert
+            class1.Should().BeEquivalentTo(class2,
+                opts => opts.Excluding(o => o.Field3).Excluding(o => o.Property1));
         }
 
         [Fact]
@@ -303,55 +296,49 @@ public partial class SelectionRulesSpecs
                 () => class1.Should().BeEquivalentTo(class2, opts => opts.Excluding(o => o.Property1));
 
             // Assert
-            act.Should().Throw<XunitException>().WithMessage("Expected*Field3*");
+            act.Should().Throw<XunitException>()
+                .WithMessage("Expected*Field3*");
         }
 
         [Fact]
-        [SuppressMessage("ReSharper", "StringLiteralTypo")]
         public void When_excluding_member_names_it_should_pass_if_only_the_excluded_members_are_different()
         {
             // Arrange
             var class1 = new ClassWithSomeFieldsAndProperties
             {
-                Field1 = "Lorem",
-                Field2 = "ipsum",
-                Field3 = "dolor",
-                Property1 = "sit"
+                Field1 = "First",
+                Field2 = "Second",
+                Field3 = "Rhird",
+                Property1 = "FirstProperty",
             };
 
             var class2 = new ClassWithSomeFieldsAndProperties
             {
-                Field1 = "Lorem",
-                Field2 = "ipsum"
+                Field1 = "First",
+                Field2 = "Second",
             };
 
-            // Act
-            Action act =
-                () =>
-                    class1.Should().BeEquivalentTo(class2,
-                        opts => opts.ExcludingMembersNamed("Field3", "Property1"));
-
-            // Assert
-            act.Should().NotThrow("the non-excluded fields have the same value");
+            // Act / Assert
+            class1.Should().BeEquivalentTo(class2,
+                opts => opts.ExcludingMembersNamed("Field3", "Property1"));
         }
 
         [Fact]
-        [SuppressMessage("ReSharper", "StringLiteralTypo")]
         public void When_excluding_member_names_it_should_fail_if_any_non_excluded_members_are_different()
         {
             // Arrange
             var class1 = new ClassWithSomeFieldsAndProperties
             {
-                Field1 = "Lorem",
-                Field2 = "ipsum",
-                Field3 = "dolor",
-                Property1 = "sit"
+                Field1 = "First",
+                Field2 = "Second",
+                Field3 = "Rhird",
+                Property1 = "FirstProperty",
             };
 
             var class2 = new ClassWithSomeFieldsAndProperties
             {
-                Field1 = "Lorem",
-                Field2 = "ipsum"
+                Field1 = "First",
+                Field2 = "Second",
             };
 
             // Act
@@ -359,7 +346,8 @@ public partial class SelectionRulesSpecs
                 () => class1.Should().BeEquivalentTo(class2, opts => opts.ExcludingMembersNamed("Property1"));
 
             // Assert
-            act.Should().Throw<XunitException>().WithMessage("Expected*Field3*");
+            act.Should().Throw<XunitException>()
+                .WithMessage("Expected*Field3*");
         }
 
         [Fact]
@@ -554,7 +542,8 @@ public partial class SelectionRulesSpecs
                 () => class1.Should().BeEquivalentTo(class2, opts => opts.ExcludingProperties());
 
             // Assert
-            act.Should().Throw<XunitException>().WithMessage("*dolor*color*");
+            act.Should().Throw<XunitException>()
+                .WithMessage("*dolor*color*");
         }
 
         [Fact]
@@ -583,7 +572,8 @@ public partial class SelectionRulesSpecs
                 () => class1.Should().BeEquivalentTo(class2, opts => opts.ExcludingFields());
 
             // Assert
-            act.Should().Throw<XunitException>().WithMessage("*Property3*consectetur*");
+            act.Should().Throw<XunitException>()
+                .WithMessage("*Property3*consectetur*");
         }
 
         [Fact]
@@ -858,7 +848,8 @@ public partial class SelectionRulesSpecs
             Action act = () => obj1.Should().BeEquivalentTo(obj2, opt => opt
                 .Excluding(o => o.AbstractProperty + "B"));
 
-            act.Should().Throw<ArgumentException>().WithMessage("*(o.AbstractProperty + \"B\")*cannot be used to select a member*");
+            act.Should().Throw<ArgumentException>()
+                .WithMessage("*(o.AbstractProperty + \"B\")*cannot be used to select a member*");
         }
 
 #if NETCOREAPP3_0_OR_GREATER
@@ -1327,7 +1318,8 @@ public partial class SelectionRulesSpecs
                 .Exclude(p => new { }));
 
             // Assert
-            act.Should().Throw<XunitException>().WithMessage("*Pets[1].Name*Pets[1].Age*");
+            act.Should().Throw<XunitException>()
+                .WithMessage("*Pets[1].Name*Pets[1].Age*");
         }
     }
 }